MapModal: move to components/, add drawLine pass-through
The Stats Location card was the only caller when #278 introduced this
modal, so it lived under Stats/. Month and Year views will now mount the
same modal from their Navigation bar, with `drawLine` enabled to keep
the chronological polyline the old footer maps drew. Lift the file to
`components/MapModal.tsx` so both Stats and the Month/Year navs reach
it via the same import depth.

No functional change to the Stats Location card — same chrome, same
title, same modal behaviour.
